Drop circuit breaker module, extract Configurable
